battle_waza: fn_801DA94C 96→100% (s32→u8 result var to force clrlwi)

+++ src/game/battle/battle_waza.c
@@ -1229,7 +1229,7 @@ void fn_801DA914(void* obj) {
  */
 s32 fn_801DA94C(void* obj) {
     void* resolved;
-    s32 result;
+    u8 result;
     if (obj == NULL) {
         return 0;
     }
